MOVE: notification module -> old module
