+++ src/main/java/peer/backend/repository/noti/NotificationSubscriptionRepository.java
@@ -1,8 +1,7 @@
 package peer.backend.repository.noti;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import peer.backend.entity.noti.Notification;
-import peer.backend.entity.noti.NotificationSubscription;
+import peer.backend.entity.noti.old.NotificationSubscription;
 
 public interface NotificationSubscriptionRepository extends JpaRepository<NotificationSubscription, Long> {
 }
