+++ src/main/java/peer/backend/repository/noti/NotificationTargetRepository.java
@@ -1,7 +1,7 @@
 package peer.backend.repository.noti;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import peer.backend.entity.noti.NotificationTarget;
+import peer.backend.entity.noti.old.NotificationTarget;
 
 public interface NotificationTargetRepository extends JpaRepository<NotificationTarget, Long> {
